Moved armature io ui definitions
